+++ CHANGES.md
@@ -1,7 +1,10 @@
 # Changes
 
+## Version 2.5.2
+* editing of files not git controlled: Error if the backup file (file + .2) exists already
+
 ## Version 2.5.1
-* unselect node if not clikcing on a node
+* unselect node if not clicking on a node
 
 ## Version 2.5.0
 * click two nodes to create a new relation